+++ estimate-server/src/server.rs
@@ -22,7 +22,7 @@ pub fn calculate_taxes(input_info: TaxInfo) -> EstimaterResult<TaxResults> {
 
     let federal_tax = intermediate.taxable_income * (input_info.federal_tax_rate_percent / 100.0);
     let state_tax = intermediate.taxable_income * (input_info.state_tax_rate_percent / 100.0);
-    let net_income = (input_info.gross_yearly_income as f32)
+    let net_income = (input_info.gross_yearly_income as f64)
         - -(federal_tax * intermediate.taxable_income)
         - (state_tax * intermediate.taxable_income);
     Ok(TaxResults::new(federal_tax, state_tax, net_income))
@@ -30,12 +30,12 @@ pub fn calculate_taxes(input_info: TaxInfo) -> EstimaterResult<TaxResults> {
 
 /// Represents data / results generated mid calculation that get reused.
 struct IntermediateTaxData {
-    taxable_income: f32,
+    taxable_income: f64,
 }
 
 impl IntermediateTaxData {
     pub(crate) fn new(input_info: &TaxInfo) -> Self {
-        let taxable_income = input_info.gross_yearly_income as f32 - input_info.pre_tax_deducations;
+        let taxable_income = input_info.gross_yearly_income as f64 - input_info.pre_tax_deducations;
         Self { taxable_income }
     }
 }
